fs: 64-bit files and media -- >4 GiB on exFAT, 4Kn sectors, LBAs past 2 TiB (#676, #683)
The 32->64-bit widening, both halves in one change because they are the same
assumption removed twice: "everything fits 32 bits".

Files (#676): the whole length model -- ra8_fs_file_t size/offset/valid,
seek/tell/size/truncate, stat, the listdir callback -- is 64-bit, matching
exFAT DataLength/ValidDataLength on disk. The writable-set guard that refused
a non-zero DataLength high word is gone; the stream flush, truncate, unlink
and check paths read and write the full 64-bit lengths (the unlink path read
only the low word and would have leaked most of a >4 GiB file). FAT12/16/32
keep the format 4 GiB-1 ceiling (DIR_FileSize is 32-bit), enforced with
invalid_size at the FAT write and truncate boundaries instead of wrapping.

Media (#683): the backend interface (read/write/get_capacity/erase), the
mount LBA fields, the GPT parser and every cluster computation carry 64-bit
LBAs, so partitions past 2 TiB select and mount instead of returning
not_supported. The sector size is the runtime property it always claimed to
be: the backend block size (512/1024/2048/4096) is probed at mount,
cross-checked against the BPB/VBR (a 512e image on a 4Kn device is refused,
not reinterpreted), and both formatters lay geometry at any of the four
sizes. Sector buffers size to a 4 KiB max; the GPT machinery moved to its own
TU and the byte codec to its own header when mount.c and protos_b crossed the
source cap.

SIMULATION-VERIFIED, per the issue terms -- no 4Kn or multi-TiB medium
exists on the bench, and the header, README and tests say so:
- test_ra8_fs_exfat_large_file: a >4 GiB file on a 6 GiB sparse fake
  (windowed store, no RAM cost) -- truncate-grow across the line, seek/tell/
  read at 0xFFFFFFFF +/- 1, real bytes appended past 4 GiB and spanning the
  boundary, truncate down through it; fsck.exfat -n clean.
- test_ra8_fs_4kn: FAT16/FAT32/exFAT format+mount+round-trip at 4096-byte
  sectors, ra8_fs_check clean, both mismatch refusals; fsck.exfat -n clean.
- test_ra8_fs_large_media: an exFAT volume planted at absolute LBA 2^32 on a
  3 TiB sparse fake, mounted through a GPT entry whose FirstLBA tops 32 bits,
  file I/O proven to address the backend past 2 TiB.
The existing 60+ fs binaries pass unchanged (512-byte/32-bit regression
surface), and all 26 dumped exFAT images stay fsck.exfat -n clean.

Consumers updated in the same change: the ra8_io blockdev bridge and the
SD-SPI / USB-MSC backends range-check their own 32-bit transports instead of
truncating, ra8_io_vfs stat reports 64-bit sizes, and the shelf/import
examples skip-not-wrap oversized entries. MISRA 10.6/10.8 composites fixed at
the root; the 15.5 deltas are the dispositioned D-001 early-return deviation.

+++ examples/ek_ra8d2/hw_pending/import_reader/main.c
@@ -513,13 +513,13 @@ imp_read_cache(ra8_fs_mount_t* mount, const char* path, uint32_t* out_len)
   if (err != k_ra8_ok) {
     return err;
   }
-  uint32_t size = 0U;
+  uint64_t size = 0U;
   err           = ra8_fs_size(file, &size);
   if (err != k_ra8_ok) {
     (void)ra8_fs_close(file);
     return err;
   }
-  if (size > (uint32_t)sizeof(s_imp_readback)) {
+  if (size > (uint64_t)sizeof(s_imp_readback)) {
     (void)ra8_fs_close(file);
     return k_ra8_err_no_mem;
   }
