+++ tests/support/fs_exfat_stream_test_util.h
@@ -240,7 +240,7 @@ stream_expect_contents(ra8_fs_mount_t* h, const char* name, uint32_t total, uint
 {
   ra8_fs_file_t* f = nullptr;
   TEST_ASSERT_EQ(k_ra8_ok, ra8_fs_open(h, name, k_ra8_fs_mode_read, &f));
-  uint32_t size = 0U;
+  uint64_t size = 0U;
   TEST_ASSERT_EQ(k_ra8_ok, ra8_fs_size(f, &size));
   TEST_ASSERT_EQ(total, size);
   static uint8_t s_buf[k_xs_big_chunk];
